+++ spec/spec_helper.rb
@@ -2,6 +2,12 @@
 
 require "offense_to_corrector"
 
+module AstHelpers
+  def s(type, *children)
+    RuboCop::AST::Node.new(type, children)
+  end
+end
+
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
@@ -12,4 +18,6 @@ RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = :expect
   end
+
+  config.include AstHelpers
 end